Prepare campaign for C23_WMDE_Mobile_DE_07
Based on the VAR of 06

Ticket: https://phabricator.wikimedia.org/T350457
