+++ banners/mobile/form_items.ts
@@ -12,7 +12,7 @@ export function createFormItems( translations: Translator, amountFormatter: Numb
 			Intervals.MONTHLY,
 			Intervals.YEARLY
 		)
-		.setAmounts( 5, 15, 25, 50, 100 )
+		.setAmounts( 10, 15, 25, 50, 100 )
 		.setPaymentMethods(
 			PaymentMethods.PAYPAL,
 			PaymentMethods.DIRECT_DEBIT,
