--- banners/mobile/styles/MiniBannerVar.scss
@@ -1,158 +1,0 @@
-@use '@src/themes/Mikings/variables/colors';
-@use '@src/themes/Mikings/variables/globals';
-@use 'sass:color';
-
-$height: 288px !default;
-$green: #6e9e00;
-
-.wmde-banner {
-	&-mini {
-		display: flex;
-		flex-direction: column;
-		min-height: $height;
-		padding: 16px 0;
-		position: relative;
-		border: 2px solid colors.$primary;
-		background: colors.$white;
-
-		&-close {
-			position: absolute;
-			height: 36px;
-			width: 36px;
-			top: 11px;
-			right: 16px;
-			text-align: center;
-			background: colors.$white;
-			padding: 10px;
-			z-index: 2;
-
-			&-button {
-				margin-top: auto;
-				float: right;
-				height: 16px;
-				line-height: 16px;
-				width: 16px;
-				background: colors.$white;
-				z-index: 2;
-
-				svg {
-					height: 16px;
-					width: 16px;
-				}
-
-				&:hover {
-					cursor: pointer;
-				}
-			}
-		}
-
-		&-headline {
-			height: 25px;
-			text-align: center;
-			margin: 0 16px 16px;
-
-			&-background {
-				position: relative;
-				text-align: left;
-
-				@media ( min-width: 400px ) {
-					text-align: center;
-				}
-
-				/* single line above container */
-				&::before {
-					content: '';
-					display: block;
-					background: colors.$primary;
-					width: 100%;
-					height: 1px;
-					position: absolute;
-					top: 50%;
-					z-index: 1;
-				}
-			}
-
-			&-content {
-				position: relative;
-				display: inline-block;
-				font-weight: bold;
-				font-size: 14px;
-				line-height: 25px;
-				color: colors.$black;
-				background: colors.$white;
-				padding: 0 5px;
-				z-index: 2;
-
-				@media ( min-width: 330px ) {
-					font-size: 16px;
-				}
-
-				@media ( min-width: 360px ) {
-					font-size: 18px;
-				}
-			}
-		}
-
-		&-slideshow {
-			display: flex;
-			flex-direction: column;
-			flex: 1 1 auto;
-		}
-
-		&-button-group {
-			display: flex;
-			justify-content: center;
-		}
-
-		&-button,
-		&-button-preselect {
-			width: 50%;
-			height: 40px;
-			line-height: 40px;
-			border-radius: 20px;
-			font-weight: bold;
-			color: colors.$white;
-			margin: 0 16px;
-			font-size: 14px;
-			white-space: nowrap;
-
-			@media ( min-width: 370px ) {
-				font-size: 16px;
-			}
-		}
-
-		&-button {
-			background: colors.$secondary;
-
-			&:hover,
-			&:focus {
-				background: colors.$secondary-hover;
-			}
-		}
-
-		&-button-preselect {
-			background: $green;
-
-			&:hover,
-			&:focus {
-				background: color.adjust( $green, $lightness: -5% );
-			}
-		}
-
-		.smallprint-mini {
-			text-align: center;
-			font-size: 11px;
-			margin-top: 12px;
-			margin-bottom: -5px;
-
-			a {
-				color: colors.$gray;
-
-				&:hover,
-				&:focus {
-					text-decoration: underline;
-				}
-			}
-		}
-	}
-}
