--- banners/mobile/styles/styles_var.scss
@@ -1,18 +1,0 @@
-@use 'src/components/BannerConductor/banner-transition';
-@use 'src/themes/UseOfFunds/UseOfFunds';
-@use 'src/themes/Mikings/defaults';
-@use './Banner';
-@use './MiniBannerVar' with (
-	$height: 288px
-);
-@use './FullPageBanner';
-@use 'src/themes/Mikings/Footer/Footer';
-@use 'src/themes/Mikings/Footer/SelectionInput';
-@use 'src/themes/Mikings/DonationForm/MultiStepDonation';
-@use 'src/themes/Mikings/DonationForm/Forms/UpgradeToYearlyButtonForm';
-@use 'src/themes/Mikings/DonationForm/SubComponents/SelectGroup';
-@use 'src/themes/Mikings/DonationForm/SubComponents/SelectCustomAmount';
-@use 'src/themes/Mikings/DonationForm/SubComponents/SmsBox';
-@use 'src/themes/Mikings/Slider/Slider';
-@use 'src/themes/Mikings/SoftClose/SoftClose';
-@use 'src/themes/Mikings/ProgressBar/ProgressBar';
