+++ src/themes/Mikings/variables/_colors.scss
@@ -4,6 +4,7 @@ $secondary: #da9000;
 $secondary-light: #caac72;
 $secondary-hover: #bd7d00;
 $error: #dd3333;
+$green: #6e9e00;
 
 $white: #ffffff;
 $black: #000000;
