--- test/banners/mobile/components/BannerVar.spec.ts
@@ -1,159 +1,0 @@
-import { afterEach, beforeEach, describe, test, vi } from 'vitest';
-import { mount, VueWrapper } from '@vue/test-utils';
-import Banner from '../../../../banners/mobile/components/BannerVar.vue';
-import { BannerStates } from '@src/components/BannerConductor/StateMachine/BannerStates';
-import { PageScroller } from '@src/utils/PageScroller/PageScroller';
-import { useOfFundsContent } from '@test/banners/useOfFundsContent';
-import { newDynamicContent } from '@test/banners/dynamicCampaignContent';
-import { CurrencyEn } from '@src/utils/DynamicContent/formatters/CurrencyEn';
-import { formItems } from '@test/banners/formItems';
-import { softCloseFeatures } from '@test/features/SoftCloseMobile';
-import { useOfFundsFeatures, useOfFundsScrollFeatures } from '@test/features/UseOfFunds';
-import { miniBannerFeatures } from '@test/features/MiniBanner';
-import { donationFormFeatures } from '@test/features/forms/MainDonation_UpgradeToYearlyButton';
-import { useFormModel } from '@src/components/composables/useFormModel';
-import { resetFormModel } from '@test/resetFormModel';
-import { DynamicContent } from '@src/utils/DynamicContent/DynamicContent';
-import { bannerContentAnimatedTextFeatures } from '@test/features/BannerContent';
-import { fullPageBannerFeatures } from '@test/features/FullPageBanner';
-import { miniBannerPreselectFeatures } from '@test/features/MiniBannerPreselect';
-import { Tracker } from '@src/tracking/Tracker';
-
-let pageScroller: PageScroller;
-let tracker: Tracker;
-const formModel = useFormModel();
-const translator = ( key: string ): string => key;
-describe( 'BannerVar.vue', () => {
-
-	let wrapper: VueWrapper<any>;
-	beforeEach( () => {
-		resetFormModel( formModel );
-		vi.useFakeTimers();
-
-		pageScroller = {
-			scrollIntoView: vi.fn(),
-			scrollToTop: vi.fn()
-		};
-
-		tracker = {
-			trackEvent: vi.fn()
-		};
-	} );
-
-	const getWrapper = ( dynamicContent: DynamicContent = null ): VueWrapper<any> => {
-		// attachTo the document body to fix an issue with Vue Test Utils where
-		// clicking a submit button in a form does not fire the submit event
-		wrapper = mount( Banner, {
-			attachTo: document.body,
-			props: {
-				bannerState: BannerStates.Pending,
-				useOfFundsContent,
-				pageScroller
-			},
-			global: {
-				mocks: {
-					$translate: translator
-				},
-				provide: {
-					translator: { translate: translator },
-					dynamicCampaignText: dynamicContent ?? newDynamicContent(),
-					formActions: { donateWithAddressAction: 'https://example.com', donateWithoutAddressAction: 'https://example.com' },
-					currencyFormatter: new CurrencyEn(),
-					formItems,
-					tracker
-				}
-			}
-		} );
-
-		return wrapper;
-	};
-
-	afterEach( () => {
-		wrapper.unmount();
-		vi.restoreAllMocks();
-		vi.useRealTimers();
-	} );
-
-	// skipped because the sentence is not part of the current test
-	describe.skip( 'Content', () => {
-		test.each( [
-			[ 'expectHidesAnimatedVisitorsVsDonorsSentenceInMessage' ],
-			[ 'expectShowsAnimatedVisitorsVsDonorsSentenceInMessage' ],
-			[ 'expectHidesAnimatedVisitorsVsDonorsSentenceInSlideShow' ],
-			[ 'expectShowsAnimatedVisitorsVsDonorsSentenceInSlideShow' ]
-		] )( '%s', async ( testName: string ) => {
-			await bannerContentAnimatedTextFeatures[ testName ]( getWrapper );
-		} );
-	} );
-
-	describe( 'Donation Form Happy Paths', () => {
-		test.each( [
-			[ 'expectMainDonationFormSubmitsWhenSofortIsSelected' ],
-			[ 'expectMainDonationFormSubmitsWhenYearlyIsSelected' ],
-			[ 'expectMainDonationFormGoesToUpgrade' ],
-			[ 'expectUpgradeToYearlyFormSubmitsUpgrade' ],
-			[ 'expectUpgradeToYearlyFormSubmitsDontUpgrade' ],
-			[ 'expectUpgradeToYearlyFormGoesToMainDonation' ]
-		] )( '%s', async ( testName: string ) => {
-			await donationFormFeatures[ testName ]( getWrapper() );
-		} );
-	} );
-
-	describe( 'Soft Close', () => {
-		test.each( [
-			[ 'expectShowsSoftCloseOnMiniBannerClose' ],
-			[ 'expectDoesNotShowSoftCloseOnFullBannerClose' ],
-			[ 'expectEmitsSoftCloseCloseEvent' ],
-			[ 'expectEmitsSoftCloseMaybeLaterEvent' ],
-			[ 'expectEmitsSoftCloseTimeOutEvent' ],
-			[ 'expectEmitsBannerContentChangedOnSoftClose' ]
-		] )( '%s', async ( testName: string ) => {
-			await softCloseFeatures[ testName ]( getWrapper() );
-		} );
-	} );
-
-	describe( 'Use of Funds', () => {
-		test.each( [
-			[ 'expectShowsUseOfFunds' ],
-			[ 'expectHidesUseOfFunds' ]
-		] )( '%s', async ( testName: string ) => {
-			await useOfFundsFeatures[ testName ]( getWrapper() );
-		} );
-
-		test.each( [
-			[ 'expectScrollsToFormWhenCallToActionIsClicked' ],
-			[ 'expectScrollsToLinkWhenCloseIsClicked' ]
-		] )( '%s', async ( testName: string ) => {
-			await useOfFundsScrollFeatures[ testName ]( getWrapper(), pageScroller );
-		} );
-	} );
-
-	describe( 'Mini Banner', () => {
-		test.each( [
-			[ 'expectSlideShowPlaysWhenMiniBannerBecomesVisible' ],
-			[ 'expectSlideShowStopsWhenFullBannerBecomesVisible' ],
-			[ 'expectShowsFullPageWhenCallToActionIsClicked' ],
-			[ 'expectEmitsBannerContentChangedEventWhenCallToActionIsClicked' ]
-		] )( '%s', async ( testName: string ) => {
-			await miniBannerFeatures[ testName ]( getWrapper() );
-		} );
-
-		test.each( [
-			[ 'expectShowsFullPageWhenPreselectIsClicked' ],
-			[ 'expectPreselectsAmountWhenPreselectIsClicked' ],
-			[ 'expectTrackingEventIsFiredWhenPreselectIsClicked' ]
-		] )( '%s', async ( testName: string ) => {
-			await miniBannerPreselectFeatures[ testName ]( getWrapper(), tracker );
-		} );
-
-	} );
-
-	describe( 'Full Page Banner', () => {
-		test.each( [
-			[ 'expectEmitsCloseEvent' ]
-		] )( '%s', async ( testName: string ) => {
-			await fullPageBannerFeatures[ testName ]( getWrapper() );
-		} );
-	} );
-
-} );
